+++ src/components/hero/campaignHeroSubText.tsx
@@ -3,9 +3,11 @@ import {
   HorizontalRule,
   Text,
   Spacer,
+  Box,
+  Tooltip,
 } from "@nypl/design-system-react-components";
 
-const CampaignHeroSubText = () => {
+const CampaignHeroSubText = ({ featuredItem }: any) => {
   return (
     <>
       {/* To Do: make the link color blue: */}
@@ -29,29 +31,33 @@ const CampaignHeroSubText = () => {
           }}
           href="https://digitalcollections.nypl.org/about#nypl_harmful_content_statement"
           aria-label="Learn more about harmful content"
+          target="_blank"
         >
           Learn more
         </DSLink>
       </Text>
       <HorizontalRule />
-      <Text color="ui.typography.body" mb="0px">
-        Featured Image:{" "}
-        <DSLink
-          color="var(--nypl-colors-ui-link-primary) !important"
-          __css={{
-            _hover: {
-              color: "var(--nypl-colors-ui-link-secondary) !important",
-            },
-            _visited: {
-              color: "var(--nypl-colors-ui-link-tertiary) !important",
-            },
-          }}
-          href="https://digitalcollections.nypl.org/items/510d47e0-cb17-a3d9-e040-e00a18064a99"
-        >
-          {" "}
-          Momoyogusa = Flowers of a Hundred Generations.{" "}
-        </DSLink>
-      </Text>
+      <Box>
+        <Text color="ui.typography.body" mb="0px" noOfLines={2}>
+          Featured Image:
+          <DSLink
+            color="var(--nypl-colors-ui-link-primary) !important"
+            __css={{
+              display: "inline !important",
+              _hover: {
+                color: "var(--nypl-colors-ui-link-secondary) !important",
+              },
+              _visited: {
+                color: "var(--nypl-colors-ui-link-tertiary) !important",
+              },
+            }}
+            href={featuredItem.href}
+          >
+            {" "}
+            {featuredItem.title}{" "}
+          </DSLink>
+        </Text>
+      </Box>
     </>
   );
 };