+++ src/pages/index.tsx
@@ -1,29 +1,93 @@
 import CampaignHero from "../components/hero/campaignHero";
-import { featuredImageID } from "../utils/utils";
+import { TemplateAppContainer } from "@nypl/design-system-react-components";
+import data from "@/data/lanes";
 import {
-  SkipNavigation,
-  TemplateAppContainer,
-} from "@nypl/design-system-react-components";
+  getNumItems,
+  getNumDigitizedItems,
+  featuredImageID,
+  getAPIUri,
+  apiCall,
+  getItemDataFromImageID,
+} from "@/utils/utils";
+import Header from "@/components/header/header";
+import HomePageMainContent from "@/components/homePageMainContent/homePageMainContent";
+import ExploreFurther from "@/components/exploreFurther/exploreFurther";
+import appConfig from "appConfig";
+import { imageURL } from "@/utils/utils";
+import NotificationBanner from "@/components/notificationBanner/notificationBanner";
 
 export default function Home(props: any) {
   return (
-    <TemplateAppContainer
-      aboveHeader={<p> Notification banner </p>}
-      header={<p> Header </p>}
-      breakout={<CampaignHero featuredImageID={props.featuredImageID} />}
-      contentPrimary={
-        <>
-          <p>First swim lane</p>
-          <p>Featured Content</p>
-          <p>Rest of swim lanes</p>
-          <p>Explore further links</p>
-        </>
-      }
-      renderSkipNavigation={true}
-    />
+    <>
+      {/**
+       * * @TODO: Header will need to be pulled into a reusable Layout component (DC Facelift phase 2)
+       * * Let this be @7emansell 's problem if possible **/}
+      <NotificationBanner />
+      <Header />
+      <TemplateAppContainer
+        breakout={
+          <CampaignHero
+            featuredItem={props.featuredItem}
+            numberOfDigitizedItems={props.numberOfDigitizedItems}
+          />
+        }
+        contentPrimary={
+          <HomePageMainContent
+            randomNumber={props.randomNumber}
+            lanesWithNumItems={props.lanesWithNumItems}
+          />
+        }
+      />
+      <ExploreFurther />
+    </>
   );
 }
 
-export async function getServerSideProps() {
-  return { props: { featuredImageID: featuredImageID() } };
+export async function getServerSideProps(context: any) {
+  const lanes = data.lanes;
+  const flatCollections = [].concat(...lanes.map((lane) => lane.collections));
+  const collectionsWithNumItems = await Promise.allSettled(
+    flatCollections.map(async (collection) => {
+      try {
+        const numItems = await getNumItems(collection.uuid);
+        return { ...collection, numItems };
+      } catch (error) {
+        return { ...collection, numItems: 0 };
+      }
+    })
+  );
+  const updatedLanes = lanes.map((lane) => {
+    const updatedCollections = lane.collections.map(() => {
+      const result = collectionsWithNumItems.shift();
+      return result.status === "fulfilled"
+        ? result.value
+        : { ...result, value: {} };
+    });
+    return { ...lane, collections: updatedCollections };
+  });
+  const randomNumber = Math.floor(Math.random() * 2);
+
+  //pass query param to featuredImageID function to check if it is legit
+  const imageID = context.query.imageID
+    ? featuredImageID(context.query.imageID)
+    : featuredImageID();
+
+  const dataFromUri = await getItemDataFromImageID(imageID);
+  const numDigitizedItems = await getNumDigitizedItems();
+  const featuredItemObject = {
+    imageID: imageID,
+    imageSrc: imageURL(imageID),
+    uuid: dataFromUri.uuid,
+    title: dataFromUri.title,
+    href: `${appConfig.DC_URL}/items/${dataFromUri.uuid}`,
+  };
+
+  return {
+    props: {
+      randomNumber,
+      lanesWithNumItems: updatedLanes,
+      featuredItem: featuredItemObject,
+      numberOfDigitizedItems: numDigitizedItems,
+    },
+  };
 }